Added properties Mouse.IsRelative property and MouseState.DeltaX/Y.
In XNA relative mouse input is usually created by resetting the mouse position in each frame. This method is not accurate for high framerates because a fraction of the mouse movement is lost. This method also does not work if Mouse.SetPosition() does not work.
The correct, accurate solution for relative mouse input in Win/DX is using the raw mouse device. On Win8 it is the OnMouseMoved event.
The new code also works when the old method (calling SetPosition() each frame) is used on other platforms.